Eliminate 'gems' dependency
It was pretty silly, in retrospect, to get JSON from the RubyGems API,
and then have Gems convert it into a Ruby object, and then have a Ruby
script convert it back into JSON, and then have UPM unmarshal it into
a struct, and then marshal it back into JSON, and then have Goval
unmarshal it into another struct, and then marshal it into Protobuf,
and then have the web frontend unmarshal it into a map before it
actually gets displayed.

That's interop for you, I guess, but at least we can get rid of the
first step.