Fixed the support for http/https imports in nodejs
We would parse http/https imports as NPM import through UPM, which would cause NPM to trigger errors as it tries to import packages using the URL path.

Nodejs is on its way to support URL import paths and the browser already does it (Which is used by CDNs and tools like skypack), so this change should allow support for both future version of node and imports for the browser.

Also added a test suite for this backend and a make target to run the tests on demand.