Add support for fetching the custom version through the API
When the dynamic version code was executed through niv, the .git folder would be missing. This is because niv downloads the archive from github using an URL without getting the .git folder, since it is never really pushed.

That's a big problem for us and this PR fixes that issue by adding an if and checking if said .git folder exists. If not, it will use the version information from the API and parse the short sha from there.