Normalize List results in test-suite as well
It shouldn't really matter here, but in case we get any that come through unformatted.

What I would like to do is expose b.NormalizePackageName as a string -> api.PkgName, so
if we have a PkgName we have some guarantees around it, but I don't have a clear way to do that yet.